requirements.txt
fastapi[standard]
SQLAlchemy
passlib
python-jose
psycopg2-binary
itsdangerous
scrapy
scrapy-playwright
beautifulsoup4
tqdm